[Issue-72] 친구관리 - 검색 화면과 tab 화면을 fragment로 분리 및 생성
