+++ Picon/app/src/main/res/layout/manage_friend_activity.xml
@@ -90,25 +90,14 @@
             app:layout_constraintTop_toTopOf="@+id/manage_friend_search_et"
             app:srcCompat="@drawable/ic_search_delete" />
 
-        <com.google.android.material.tabs.TabLayout
-            android:id="@+id/manage_friend_tab_layout"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:tabTextColor="@color/brown_grey_two"
-            app:tabSelectedTextColor="@color/cornflower"
-            app:tabIndicatorColor="#5B8FF5"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/manage_friend_search_et"/>
-
-        <androidx.viewpager2.widget.ViewPager2
-            android:id="@+id/manage_friend_view_pager"
+        <FrameLayout
+            android:id="@+id/manage_friend_frame_layout"
             android:layout_width="match_parent"
             android:layout_height="0dp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/manage_friend_tab_layout" />
+            app:layout_constraintTop_toBottomOf="@+id/manage_friend_search_et" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
